requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
motor
pymongo
redis
numpy
tiktoken
google-genai
langgraph
langchain-core
mangum
boto3
ruff
mypy
pytest
pytest-asyncio
pytest-cov
httpx
python-multipart
